+++ src/attributes.rs
@@ -1873,7 +1873,7 @@ impl ParticleLayout {
     ///     .append(Attribute::POSITION)
     ///     .append(Attribute::SIZE)
     ///     .build();
-    /// let size_offset = layout.offset(Attribute::SIZE);
+    /// let size_offset = layout.byte_offset(Attribute::SIZE);
     /// assert_eq!(size_offset, Some(12));
     /// ```
     pub fn byte_offset(&self, attribute: Attribute) -> Option<u32> {